requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
djangorestframework==3.13.1
mysql-connector-python==8.0.26
protobuf==3.19.4
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
